+++ stress_test/robustness_C.py
@@ -93,5 +93,7 @@ if __name__ == "__main__":
         run_zombie_test()
     finally:
         os.remove(DB_PATH)
-        os.remove(DB_PATH + "-wal")
-        os.remove(DB_PATH + "-shm")
+        if os.path.exists(DB_PATH + "-wal"):
+            os.remove(DB_PATH + "-wal")
+        if os.path.exists(DB_PATH + "-shm"):
+            os.remove(DB_PATH + "-shm")
